huggingface-space/requirements.txt
torch
git+https://github.com/huggingface/diffusers
transformers
accelerate
sentencepiece
numpy
Pillow